Refactor add teacher and student in relation to add person
