+++ src/app.rb
@@ -78,33 +78,6 @@ class App
     @menu.option_list
   end
 
-  def add_student
-    print 'Enter the name of the student: '
-    name = gets.chomp
-    print 'Enter the age of the student: '
-    age = gets.chomp.to_i
-    print 'Has parent permission? [Y/N]: '
-    parent_permission = gets.chomp.to_s.downcase == 'y'
-    student = Student.new(age, name, parent_permission: parent_permission)
-    @people << student
-
-    puts "Student with id #{student.id} successfully created"
-  end
-
-  def add_teacher
-    print 'Enter the name of the teacher: '
-    name = gets.chomp
-    print 'Enter the age of the teacher: '
-    age = gets.chomp.to_i
-    print "And teacher's specialization: "
-    specilization = gets.chomp.to_s
-    teacher = Teacher.new(specilization, age, name)
-    @people << teacher
-
-    puts "Teacher with id #{teacher.id} successfully created"
-  end
-
-
   def exit
     puts 'Goodbye...'
   end
