+++ src/create_contents/create_person.rb
@@ -1,3 +1,6 @@
+require_relative './create_teacher'
+require_relative './create_student'
+
 def create_person
     puts 'Do you want to create a student (1) or a teacher (2)? [Input the number]: '
     type = gets.chomp.to_i
